funs_dikes.py: Speedup linear interpolation with NumPy
Speed up the model by approximately a factor 2 by using NumPy.interp() instead of the custom defined function.

Fully validated against the old behaviour and SciPy.interp1d. Even makes values on the edges of bounds a bit more accurate.

Note that the function expects a sorted array as input. Do not input unsorted arrays. Therefore, note that the rating curves are sorted when the network is generated.